+++ src/muncher/settings.js
@@ -4,15 +4,6 @@ import logger from "../logger.js";
 import { getPatreonTiers } from "./utils.js";
 import { getCobalt } from "../lib/Secrets.js";
 
-function autoACDisableEffects() {
-  const AUTO_AC = utils.versionCompare(game.data.system.data.version, "1.4.0") >= 0;
-  if (AUTO_AC) {
-    game.settings.set("ddb-importer", "character-update-policy-generate-ac-armor-effects", false);
-    game.settings.set("ddb-importer", "character-update-policy-generate-base-ac", false);
-    game.settings.set("ddb-importer", "character-update-policy-generate-ac-override-effects", false);
-  }
-}
-
 export function setRecommendedCharacterActiveEffectSettings(html) {
   $(html).find("#character-import-policy-dae-copy").prop("checked", false);
   game.settings.set("ddb-importer", "character-update-policy-dae-copy", false);
@@ -26,18 +17,6 @@ export function setRecommendedCharacterActiveEffectSettings(html) {
   game.settings.set("ddb-importer", "character-update-policy-generate-ac-feature-effects", true);
   $(html).find("#character-import-policy-active-effect-copy").prop("checked", false);
   game.settings.set("ddb-importer", "character-update-policy-active-effect-copy", false);
-
-  const AUTO_AC = utils.versionCompare(game.data.system.data.version, "1.4.0") >= 0;
-  if (AUTO_AC) {
-    autoACDisableEffects();
-  } else {
-    $(html).find("#character-import-policy-generate-ac-armor-effects").prop("checked", true);
-    game.settings.set("ddb-importer", "character-update-policy-generate-ac-armor-effects", true);
-    $(html).find("#character-import-policy-generate-base-ac").prop("checked", true);
-    game.settings.set("ddb-importer", "character-update-policy-generate-base-ac", true);
-    $(html).find("#character-import-policy-generate-ac-override-effects").prop("checked", false);
-    game.settings.set("ddb-importer", "character-update-policy-generate-ac-override-effects", false);
-  }
 
   ["class", "race", "background", "feat"].forEach((type) => {
     $(html).find(`#character-import-policy-effect-${type}-spell-bonus`).prop("checked", true);
@@ -54,9 +33,6 @@ export function setRecommendedCharacterActiveEffectSettings(html) {
 }
 
 export function getCharacterImportSettings() {
-  const AUTO_AC = utils.versionCompare(game.data.system.data.version, "1.4.0") >= 0;
-  autoACDisableEffects();
-
   const importPolicies = [
     {
       name: "name",
@@ -202,7 +178,7 @@ export function getCharacterImportSettings() {
     },
   ];
 
-  const effectImportConfig1 = [
+  const effectImportConfig = [
     {
       name: "add-item-effects",
       isChecked: game.settings.get("ddb-importer", "character-update-policy-add-item-effects") && daeInstalled,
@@ -219,20 +195,6 @@ export function getCharacterImportSettings() {
         "Dynamically generate active effects for a character. Some effects are always generated, some are optional (see below).",
       enabled: daeInstalled,
     },
-  ];
-  const effectImportConfig2 = AUTO_AC
-  ? []
-  : [
-    {
-      name: "generate-ac-armor-effects",
-      isChecked: game.settings.get("ddb-importer", "character-update-policy-generate-ac-armor-effects") && daeInstalled,
-      title: "Generate Active Effects ACs for Armor",
-      description:
-        "Dynamically add AC values as dynamic effects to armor items. If you're using DAE to auto-calculate AC, you don't need this.",
-      enabled: daeInstalled && !AUTO_AC,
-    },
-  ];
-  const effectImportConfig3 = [
     {
       name: "generate-ac-feature-effects",
       isChecked:
@@ -242,29 +204,6 @@ export function getCharacterImportSettings() {
         "Use extras in DAE to Dynamically add AC values as dynamic effects to items, this might not work as expected for some AC calculations. If unticked some ac bonuses will still be generated.",
       enabled: daeInstalled,
     },
-  ];
-  const effectImportConfig4 = AUTO_AC
-  ? []
-  : [
-    {
-      name: "generate-base-ac",
-      isChecked: game.settings.get("ddb-importer", "character-update-policy-generate-base-ac"),
-      title: "Set AC to base value",
-      description:
-        "Calculate AC base to base value, e.g. 10 +dex mod/natural armor rating. Useful if you want the AC to be correct when Armor is unequipped.",
-      enabled: !AUTO_AC,
-    },
-    {
-      name: "generate-ac-override-effects",
-      isChecked:
-        game.settings.get("ddb-importer", "character-update-policy-generate-ac-override-effects") && daeInstalled,
-      title: "[Caution] Generate DAE Override ACs",
-      description:
-        "Generate possible AC combinations as dynamic effects, these are high priority effects that override other effects. Useful if you can't calculate your AC correctly using other effects.",
-      enabled: daeInstalled && !AUTO_AC,
-    },
-  ];
-  const effectImportConfig5 = [
     {
       name: "dae-effect-copy",
       isChecked: game.settings.get("ddb-importer", "character-update-policy-dae-effect-copy") && daeSRDContentAvailable,
@@ -296,8 +235,6 @@ export function getCharacterImportSettings() {
     game.settings.set("ddb-importer", "character-update-policy-dae-copy", false);
     game.settings.set("ddb-importer", "character-update-policy-dae-effect-copy", true);
   }
-
-  const effectImportConfig = [...effectImportConfig1, ...effectImportConfig2, ...effectImportConfig3, ...effectImportConfig4, ...effectImportConfig5];
 
   const effectSelectionConfig = {
     class: [
@@ -598,7 +535,6 @@ export function getCharacterImportSettings() {
 export function updateActorSettings(html, event) {
   const selection = event.currentTarget.dataset.section;
   const checked = event.currentTarget.checked;
-  const AUTO_AC = utils.versionCompare(game.data.system.data.version, "1.4.0") >= 0;
 
   logger.debug(`Updating munching-policy-${selection} to ${checked}`);
   game.settings.set("ddb-importer", "character-update-policy-" + selection, checked);
@@ -613,9 +549,6 @@ export function updateActorSettings(html, event) {
     game.settings.set("ddb-importer", "character-update-policy-add-item-effects", true);
     $(html).find("#character-import-policy-add-character-effects").prop("checked", true);
     game.settings.set("ddb-importer", "character-update-policy-add-character-effects", true);
-  } else if (!AUTO_AC && (selection === "generate-ac-armor-effects" || selection === "generate-ac-feature-effects") && checked) {
-    game.settings.set("dae", "calculateArmor", false);
-    game.settings.set("dae", "applyBaseAC", false);
   }
 }